requirements.txt
python-dotenv==1.1.1
Flask==3.1.2
openai==1.101.0
pytest==8.4.1
faster-whisper==1.2.0
